// next-app/src/data/results.types.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Spec = string[];
export type MatchNumber = number;
export type Stage = string;
export type Home = string;
export type HomeRating = number;
export type Away = string;
export type AwayRating = number;
export type HomeWinner = boolean;
export type KnownResult = boolean;
export type Matches = SolvedMatch[];
export type Name = string;
export type Score = number;
export type Players = PlayerResult[];

export interface ResultSet {
  spec: Spec;
  results: Results | null;
}
export interface Results {
  matches: Matches;
  players: Players;
}
export interface SolvedMatch {
  match_number: MatchNumber;
  stage: Stage;
  home: Home;
  home_rating: HomeRating;
  away: Away;
  away_rating: AwayRating;
  home_winner: HomeWinner;
  known_result: KnownResult;
}
export interface PlayerResult {
  name: Name;
  score: Score;
}